+++ client/encode-tag.js
@@ -1,0 +1,5 @@
+
+export default function encodeTag(data) {
+  if (typeof data !== 'string') return '';
+  return data.replaceAll(' ', '-').toLowerCase();
+};