+++ client/use-collection.js
@@ -1,10 +1,24 @@
-import {computed} from 'vue';
+import sortBy from 'lodash/sortBy.js';
+import uniq from 'lodash/uniq.js';
+
+import {computed, reactive} from 'vue';
 import {useRoute} from 'vitepress';
+import encodeTag from './encode-tag.js';
 import {data as collections} from './collections.data.js';
 
 export default function useCollection(type = undefined) {
   const route = useRoute();
+  const urlSearchParams = new URLSearchParams(window.location.search);
+
   const path = route.path;
+  const params = Object.fromEntries(urlSearchParams.entries());
+
+  function isTagSelected(q) {
+    const {tag = '', tags = ''} = params;
+    const list = [...tag.split(','), ...tags.split(',')];
+
+    return list.includes(q) || list.includes(encodeTag(q));
+  }
 
   function findCurrentIndex() {
     const result = pages.findIndex(p => p.url === route.path);
@@ -35,5 +49,39 @@ export default function useCollection(type = undefined) {
     next: nextPage.value ? {text: nextPage.value.title, link: nextPage.value.url} : undefined,
   }));
 
-  return {pages, page, nextPage, prevnext, prevPage, path};
+  // get the tagz as well
+  const tags = computed(() => sortBy(uniq(pages.map(page => page.tags).flat(Infinity))));
+
+  // and the amount of content in each
+  const tagCounts = reactive(Object.fromEntries(tags.value.map(tag => ([tag, 0]))));
+  for (const tag of pages.map(page => page.tags).flat(Infinity)) {
+    tagCounts[tag] = tagCounts[tag] ? tagCounts[tag] + 1 : 1;
+  }
+
+  // and a selected tag reactive for filtering and that sort of thing
+  const selectedTags = reactive(Object.fromEntries(tags.value.map(tag => ([tag, isTagSelected(tag)]))));
+  const selectedTagsList = computed(() => Object.entries(selectedTags)
+    .filter(pair => pair[1] === true)
+    .map(pair => pair[0]));
+
+  // helper func to see if a set of tag filtered pages has items or not, useful for showing collection sections
+  const hasItems = (items = [], tags = {}) => {
+    const tagList = Object.entries(tags).filter(pair => pair[1] === true).map(pair => pair[0]);
+    const filteredItems = items.filter(item => tagList.every(tag => item.tags.indexOf(tag) !== -1));
+    return filteredItems.length > 0;
+  };
+
+  return {
+    hasItems,
+    pages,
+    page,
+    nextPage,
+    prevnext,
+    prevPage,
+    path,
+    tagCounts,
+    tags,
+    selectedTags,
+    selectedTagsList,
+  };
 }