+++ components/VPLCollectionItems.vue
@@ -30,13 +30,14 @@
 <script setup>
 import {computed, ref, defineAsyncComponent} from 'vue';
 import {VPButton} from 'vitepress/theme';
-import VPLCollectionItem from './VPLCollectionItem.vue';
+
+import Item from './VPLCollectionItem.vue';
 
 const Article = defineAsyncComponent({
-  loader: async () => VPLCollectionItem,
+  loader: async () => Item,
 });
 
-const {items, pager, more, size} = defineProps({
+const {items, pager, more, size, tags} = defineProps({
   items: {
     type: Array,
     required: true,
@@ -53,6 +54,10 @@ const {items, pager, more, size} = defineProps({
     type: String,
     default: 'medium',
   },
+  tags: {
+    type: Object,
+    default: () => ({}),
+  },
 });
 
 // Hardcoded pager value for now
@@ -60,8 +65,15 @@ const amount = ref(pager);
 
 // normalize data and sort
 const pages = computed(() => {
-  const datePages = items.map(item => Object.assign(item, {timestamp: item.date ? item.date : item.timestamp}));
-  return datePages.sort((a, b) => a.timestamp < b.timestamp ? 1 : -1);
+  const tagList = Object.entries(tags).filter(pair => pair[1] === true).map(pair => pair[0]);
+  return items
+    .map(item => Object.assign(item, {timestamp: item.date ? item.date : item.timestamp}))
+    .filter(item => {
+      return tagList.every(tag => {
+        return item.tags.indexOf(tag) !== -1;
+      });
+    })
+    .sort((a, b) => a.timestamp < b.timestamp ? 1 : -1);
 });
 
 const adder = () => {